feat: split rule and helpers, test rule
